separacion de responsabilidades de websocket en Debate
