+++ frontend/src/debate-handler/Debate.styled.js
