+++ frontend/src/debate-handler/DebateView.js
